[Refactored for OOP]: Fixed bugs
