+++ pkg/domain/subscriber/email-dto.go
@@ -1,4 +1,4 @@
-package mail
+package subscriber
 
 type EmailDto struct {
 	Email string `form:"email" binding:"required,email"`
